backend/requirements.txt
annotated-types
a2wsgi
anyio
beautifulsoup4
certifi
cffi
charset-normalizer
click
contourpy
cryptography
cycler
Cython
distro
exceptiongroup
fastapi
fire
fonttools
h11
httpcore
httpx
idna
ImageIO
imgaug
importlib_resources
iniconfig
jiter
kiwisolver
lazy-loader
lmdb
lxml
markitdown
matplotlib
networkx
numpy
openai
opencv-contrib-python
opencv-python
opencv-python-headless
opt-einsum
packaging
paddleocr
paddlepaddle
pdfminer.six
pdfplumber
pillow
pluggy
protobuf
pyclipper
pycparser
pydantic
pydantic_core
PyMuPDF
pyparsing
pypdf
pypdfium2
pytesseract
pytest
python-dateutil
python-docx
python-dotenv
python-multipart
PyYAML
RapidFuzz
requests
safetensors
scikit-image
scipy
shapely
six
sniffio
soupsieve
starlette
termcolor
tifffile
tomli
tqdm
typing_extensions
urllib3
uvicorn
zipp
